requirements.txt
python-telegram-bot==22.1
requests==2.32.3